stock_analysis_app/requirements.txt
streamlit
yfinance
pandas
ta
google-generativeai
plotly
python-dotenv
crewai
crewai-tools
prophet
duckduckgo-search
beautifulsoup4
langchain-google-genai
scipy
numpy
